version 3.1, 3.2 and 3.2.1
3.2.1
October, 24th 2007

    Previous version modify the position of the labels in the diagrams in some cases, fixed.

    When you edited operation body through shortcut or context menu with add operation profile on body edition set, the operation profile was not be showed, fixed.

3.2
October, 21th 2007

    The SVG produced for a component diagram contained floating point numbers. Depending on the settings associated to your country, the representation of a float can uses a ',' rather than a '.', not compatible with SVG. Now all the produced numbers are integers.

    Add xml:space="preserve" in the SVG text forms, at least to indent arguments given on several lines in a sequence diagram.

    It was not possible to add a state diagram icon and an activity diagram	icon in a state diagram, component diagram, sequence diagram, collaboration diagram, object diagram, use case diagram, deployment diagram and activity diagram. Fixed.

    The interfaces provided and required by a nested subcomponent are automatically provided and required by the direct container component, but this didn't work for more levels, fixed.

    In a sequence diagram, the arguments given on several lines are shown pseudo centered, fixed.

    It was not possible for a component to require and provide the same interface, this limitation is removed.

    Modify the component dialog to set the realizing classes (wrongly labeled realized classes in previous releases) in a dedicated tab rather than to set them with the provided classes.

    Plug-out upgrade 1.10.1
    Since Bouml 2.28 the activities are not added, fixed.
    Rename the operation UmlBaseComponent::realizedClasses to realizingClasses. Warning, doesn't modify the exiting calls of realizedClasses. Rename the parameter realized of the operation UmlBaseComponent::set_AssociatedClasses to realizing
    Add operation PhpSettings::classUmlStereotype.
    Add operations UmlBaseItem::isMarked, UmlBaseItem::set_isMarked, UmlBaseItem::markedItems, UmlBaseItem::referencedBy, UmlBaseState::set_AssociatedDiagram, UmlBasePackage::loadProject, UmlBasePackage::saveProject and UmlBasePackage::isProjectModified if they are not defined.

    Html documentation Generator 2.13.1
    Plug-out empty
    Upgraded by plug-out upgrade.

    .pro Generator 2.1.3
    Add WITHPHP if at least two WITH* are defined. Note : compatible with old plug-outs without Php.

    Because the format of the BOUML files is changed, the previous releases of BOUML cannot read a project saved with this version. Obviously this release is able to read the projects made by previous releases of BOUML.

3.1
October, 14th 2007

    Since the 3.0 when you load a project Java management and default definition is forced to false if Php management and default definition is not set, fixed.

    Add force body generation flag on the operations to allow to generate their body even if preserve operation's body is set.
    Usefull to generate the implementation of the state machines in all the cases.

    Java Generator 2.13
    C++ Generator 2.10
    Fixe the indent of the first item of an enum itself embeded in an other class.
    To manage operations having their body generated even if preserve operation's body is set.

    Php Generator 1.1
    To manage operations having their body generated even if preserve operation's body is set.

    State machine generator 1.4
    To set force body generation on the produced operations.
    Upgraded by plug-out upgrade.

    Roundtrip body 1.5
    To not change the body of the operations having force body generation set.

    Global change 1.1
    To manage Php definitions.
    Upgraded by plug-out upgrade.

    Html documentation Generator 2.13
    To generate Php definitions.
    Upgraded by plug-out upgrade.

    Plug-out upgrade 1.10
    To add missing vector allocations in Java definitions.
    To add Php management and force body generation.
    Note : you must define WITHPHP when your plug-out is implemented in C++ and you defined WITHCPP WITHJAVA and WITHIDL.

    Plug-out empty
    Upgraded by plug-out upgrade.

    Because the format of the BOUML files is changed, the previous releases of BOUML cannot read a project saved with this version. Obviously this release is able to read the projects made by previous releases of BOUML.

+++ src/diagram/ArrowJunctionCanvas.cpp
@@ -121,11 +121,11 @@ void ArrowJunctionCanvas::draw(QPainter & p) {
     case UmlProvided:
       p.drawPixmap(QPoint((int) x(), (int) y()), *providedPixmap);
       if (fp != 0)
-	fprintf(fp, "<ellipse fill=\"none\" stroke=\"black\" stroke-width=\"1\" stroke-opacity=\"1\" cx=\"%d\" cy=\"%d\" rx=\"%g\" ry=\"%g\" />\n",
+	fprintf(fp, "<ellipse fill=\"none\" stroke=\"black\" stroke-width=\"1\" stroke-opacity=\"1\" cx=\"%d\" cy=\"%d\" rx=\"%d\" ry=\"%d\" />\n",
 		// 4 is the margin inside the pixmap
 		(int) (x() + PROVIDED_RADIUS + 4),
 		(int) (y() + PROVIDED_RADIUS + 4),
-		PROVIDED_RADIUS, PROVIDED_RADIUS);
+		(int) PROVIDED_RADIUS, (int) PROVIDED_RADIUS);
       break;
     default:
       // to avoid compiler warning
